fix timestamp parser
